# docs/openapi.yaml
openapi: 3.0.0
paths:
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User successfully registered. Verification email sent. No tokens returned until email is verified.
        '400':
          description: Invalid input
        '409':
          description: Email already registered
      summary: Register a new user
      tags: &ref_0
        - Authentication
  /auth/verify-email:
    post:
      operationId: AuthController_verifyEmail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailDto'
      responses:
        '200':
          description: Email verified successfully. Tokens set as HTTP-only cookies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    type: object
        '400':
          description: Invalid or expired verification code
        '429':
          description: 'Too many attempts. Rate limit: 10 per minute'
      summary: Verify email with 6-digit code (sets HTTP-only cookies)
      tags: *ref_0
  /auth/resend-verification:
    post:
      operationId: AuthController_resendVerification
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationDto'
      responses:
        '200':
          description: Verification code sent successfully
        '400':
          description: User not found or email already verified
        '429':
          description: 'Too many requests. Rate limit: 3 per minute'
      summary: Resend email verification code
      tags: *ref_0
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Successfully authenticated. Tokens set as HTTP-only cookies.
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      user:
                        type: object
                    description: Login successful - tokens set in cookies
                  - type: object
                    properties:
                      requires2FA:
                        type: boolean
                        example: true
                      tempToken:
                        type: string
                      message:
                        type: string
                    description: 2FA required - use temp token to verify
        '401':
          description: Invalid credentials
        '403':
          description: Email not verified
      summary: Login with email and password (sets HTTP-only cookies)
      tags: *ref_0
  /auth/refresh:
    post:
      description: Reads refreshToken from HTTP-only cookie and issues new tokens
      operationId: AuthController_refresh
      parameters: []
      responses:
        '200':
          description: Token refreshed successfully. New tokens set as HTTP-only cookies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tokens refreshed successfully
        '401':
          description: Invalid or missing refresh token
      summary: Refresh access token using refresh token from cookie
      tags: *ref_0
  /auth/forgot-password:
    post:
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordDto'
      responses:
        '200':
          description: Password reset code sent if email exists (message returned regardless for security)
        '429':
          description: 'Too many requests. Rate limit: 3 per minute'
      summary: Request password reset code
      tags: *ref_0
  /auth/verify-reset-code:
    post:
      operationId: AuthController_verifyResetCode
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyResetCodeDto'
      responses:
        '200':
          description: Reset code is valid
        '400':
          description: Invalid or expired reset code
        '429':
          description: 'Too many attempts. Rate limit: 5 per minute'
      summary: Verify password reset code is valid
      tags: *ref_0
  /auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordDto'
      responses:
        '200':
          description: Password reset successfully
        '400':
          description: Invalid or expired reset code
        '429':
          description: 'Too many attempts. Rate limit: 5 per minute'
      summary: Reset password with verification code
      tags: *ref_0
  /auth/set-password:
    post:
      description: Allows OAuth users (Google login) who don't have a password to set one. Once set, users can log in with either method. This endpoint ONLY works if the user has no password yet.
      operationId: AuthController_setPassword
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPasswordDto'
      responses:
        '200':
          description: Password set successfully. User can now log in with email/password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password set successfully. You can now log in with email and password.
        '400':
          description: Password is already set. Use password reset flow to change it.
        '401':
          description: Unauthorized - must be logged in
      summary: Set password for OAuth users
      tags: *ref_0
  /auth/me:
    get:
      operationId: AuthController_getProfile
      parameters: []
      responses:
        '200':
          description: User profile retrieved
        '401':
          description: Unauthorized
      summary: Get current authenticated user profile
      tags: *ref_0
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '200':
          description: Successfully logged out. Cookies cleared.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Unauthorized
      summary: Logout user by clearing authentication cookies
      tags: *ref_0
...
  /auth/google/callback:
    get:
      operationId: AuthController_googleCallback
      parameters: []
      responses:
        '302':
          description: Redirects to frontend with secure HTTP-only cookies (login) or to settings page (linking)
        '401':
          description: OAuth authentication failed
        '409':
          description: Google account already linked to another user (linking only)
      summary: Google OAuth callback (handles both login and account linking)
      tags: *ref_0
  /auth/google/link:
    get:
      operationId: AuthController_googleLink
      parameters: []
      responses:
        '302':
          description: Redirects to Google consent screen with state parameter
        '401':
          description: Unauthorized - user must be logged in
      summary: Initiate Google account linking for authenticated user
      tags: *ref_0
  /auth/account/unlink/{provider}:
    delete:
      operationId: AuthController_unlinkProvider
      parameters:
        - name: provider
          required: true
          in: path
          description: OAuth provider to unlink
          schema:
            enum:
              - google
            type: string
      responses:
        '200':
          description: Provider unlinked successfully
        '400':
          description: Cannot unlink - user has no password set or provider not linked
        '401':
          description: Unauthorized
      summary: Unlink OAuth provider from account
      tags: *ref_0
  /auth/accounts/linked:
    get:
      operationId: AuthController_getLinkedAccounts
      parameters: []
      responses:
        '200':
          description: List of linked OAuth accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                          example: google
                        linked_at:
                          type: string
                          format: date-time
                  hasPassword:
                    type: boolean
                    example: true
        '401':
          description: Unauthorized
      summary: Get list of linked OAuth providers
      tags: *ref_0
  /auth/2fa/setup:
    post:
      description: Generates TOTP secret and QR code. User must scan QR code and verify with code in step 2 (/auth/2fa/enable).
      operationId: AuthController_setup2FA
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Setup2FADto'
      responses:
        '200':
          description: QR code generated. Scan with authenticator app and verify to complete setup.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  qrCode:
                    type: string
                    description: QR code data URL for authenticator app
                  secret:
                    type: string
                    description: TOTP secret for manual entry (Base32)
        '400':
          description: 2FA already enabled
        '401':
          description: Invalid password or unauthorized
      summary: 'Setup 2FA - Step 1: Generate QR code'
      tags: *ref_0
  /auth/2fa/enable:
    post:
      description: Verifies 6-digit code from authenticator app. If valid, enables 2FA and returns backup codes.
      operationId: AuthController_enable2FA
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Enable2FADto'
      responses:
        '200':
          description: 2FA enabled successfully. Returns backup codes (save these securely).
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  backupCodes:
                    type: array
                    items:
                      type: string
                    description: One-time backup codes (save these securely)
        '400':
          description: 2FA already enabled or setup not started
        '401':
          description: Invalid 2FA code or unauthorized
      summary: 'Enable 2FA - Step 2: Verify code and activate'
      tags: *ref_0
  /auth/2fa/disable:
    post:
      operationId: AuthController_disable2FA
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Disable2FADto'
      responses:
        '200':
          description: 2FA disabled successfully. All sessions invalidated.
        '400':
          description: 2FA is not enabled
        '401':
          description: Invalid password or 2FA code
      summary: Disable 2FA for authenticated user
      tags: *ref_0
  /auth/2fa/verify:
    post:
      operationId: AuthController_verify2FA
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Verify2FADto'
      responses:
        '200':
          description: 2FA verified successfully. Tokens set as HTTP-only cookies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                  usedBackupCode:
                    type: boolean
        '401':
          description: Invalid temporary token or 2FA code
        '429':
          description: 'Too many failed 2FA attempts. Rate limit: 10 per minute'
      summary: Verify 2FA code during login (completes login flow, sets HTTP-only cookies)
      tags: *ref_0
...
  /auth/2fa/status:
    get:
      operationId: AuthController_get2FAStatus
      parameters: []
      responses:
        '200':
          description: 2FA status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  twoFactorEnabled:
                    type: boolean
                  twoFactorEnabledAt:
                    type: string
                    format: date-time
                    nullable: true
                  remainingBackupCodes:
                    type: number
        '401':
          description: Unauthorized
      summary: Get 2FA status for authenticated user
      tags: *ref_0
...
info:
  title: Crypto Watcher Bot API
  description: API for crypto trading signal bot with automated order execution
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password (minimum 8 characters, must contain uppercase, lowercase, and number or special character)
          example: SecurePass123!
          minLength: 8
        name:
          type: string
          description: User full name
          example: John Doe
      required:
        - email
        - password
    VerifyEmailDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        code:
          type: string
          description: 6-digit verification code
          example: '123456'
          minLength: 6
          maxLength: 6
      required:
        - email
        - code
    ResendVerificationDto:
      type: object
      properties:
        email:
          type: string
          description: User email address to resend verification code
          example: user@example.com
      required:
        - email
    LoginDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password
          example: password123
      required:
        - email
        - password
...